feat: 사용 중인 timeId 삭제 시 예외 발생 로직 추가

+++ src/main/java/roomescape/service/ReservationService.java
@@ -45,6 +45,10 @@ public class ReservationService {
     }
 
     public void deleteReservationTime(Long id) {
+        boolean isUsed = reservationRepository.existByTimeId(id);
+        if (isUsed) {
+            throw new IllegalArgumentException("사용 중인 time_id 입니다. 삭제가 불가합니다.");
+        }
         reservationTimeRepository.deleteById(id);
     }
 }
